Adds capability for recoding a station to a location which is related
to the original location by a colocation constraint (horizontal and
vertical uncertainty).

Addresses issue #62, but with some suggested features not yet implemented.

Fixes issue #63 (segfault when using recoding with point data)

Modifies station csv output file to include floating station statistics.

Tidies date formats in output listings to remove time component where it is 00:00:00.

+++ src/snaplib/snap/snapglob.h
@@ -83,7 +83,7 @@ SCOPE long nobs, nschp, ncon, dof;
 SCOPE int nprm;
 SCOPE double ssr, seu;
 SCOPE int iterations, converged;
-SCOPE int last_iteration_max_adjustment;
+SCOPE double last_iteration_max_adjustment;
 
 /* Other miscellaneous data */
